// clients/js-kit/src/generated/types/baseAppData.ts
/**
 * This code was AUTOGENERATED using the codama library.
 * Please DO NOT EDIT THIS FILE, instead use visitors
 * to add features, then rerun codama to update it.
 *
 * @see https://github.com/codama-idl/codama
 */

import {
  combineCodec,
  getStructDecoder,
  getStructEncoder,
  type Codec,
  type Decoder,
  type Encoder,
} from '@solana/kit';
import {
  getBasePluginAuthorityDecoder,
  getBasePluginAuthorityEncoder,
  getExternalPluginAdapterSchemaDecoder,
  getExternalPluginAdapterSchemaEncoder,
  type BasePluginAuthority,
  type BasePluginAuthorityArgs,
  type ExternalPluginAdapterSchema,
  type ExternalPluginAdapterSchemaArgs,
} from '.';

export type BaseAppData = {
  dataAuthority: BasePluginAuthority;
  schema: ExternalPluginAdapterSchema;
};

export type BaseAppDataArgs = {
  dataAuthority: BasePluginAuthorityArgs;
  schema: ExternalPluginAdapterSchemaArgs;
};

export function getBaseAppDataEncoder(): Encoder<BaseAppDataArgs> {
  return getStructEncoder([
    ['dataAuthority', getBasePluginAuthorityEncoder()],
    ['schema', getExternalPluginAdapterSchemaEncoder()],
  ]);
}

export function getBaseAppDataDecoder(): Decoder<BaseAppData> {
  return getStructDecoder([
    ['dataAuthority', getBasePluginAuthorityDecoder()],
    ['schema', getExternalPluginAdapterSchemaDecoder()],
  ]);
}

export function getBaseAppDataCodec(): Codec<BaseAppDataArgs, BaseAppData> {
  return combineCodec(getBaseAppDataEncoder(), getBaseAppDataDecoder());
}